Added entity layer. - Added PersonEntity to represent a person. - Added different domain models to represent the different types. - Added mappers in between.

+++ src/main/java/nl/orlandosmits/threekidfamily/dto/ChildDto.java
@@ -1,5 +1,5 @@
 package nl.orlandosmits.threekidfamily.dto;
 
-public record ChildDto(Integer id) {
+public record ChildDto(Long id) {
 
 }
